Update 10-26-2020
Significant progress made in last dump. It seems like if I have some other media running then the 505 server responses are less prevalent. Found a match history that I can't seem to access. It's for a TCL game. I'm hoping that this isn't true of all TCL files as that will obviously complicate this effort.