10 Minutes
Working out of "Working File for Scraping"

need to cut urls to be specific format for request to get json of information.

Need a script that cuts urls
Need to see if we can iterate over them by sending one set of cookies
Need to build way to parse each JSON file.