Header reconciliation
Need to check if I want the gameId or the matchId out of the lpl data. IT will be whichever one is in the url I think.

Also changed some of the header values in the non-lpl headers and added in the champ, gameId, and player name directly. Will have to see on a test run if that matters.