Decoding Complete?
I believe all the decoding is complete but I think I went about this the wrong way. I think the best thing to do would be to have an lpl output and a non-lpl output, then use pandas to combine them. This way I dont have to worry about items ending up in columns they shouldn't be in. I'll keep the alphabetizing stuff because I wrote it and it will only add a tenth of a second when the program is initiated.